[SPARK-49434][SPARK-49435][CONNECT][SQL] Move aggregators to sql/api
### What changes were proposed in this pull request?
This PR moves all user facing Aggregators from sql/core to sql/api.

### Why are the changes needed?
We are create a unifies Scala SQL interface. This is part of that effort.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #48267 from hvanhovell/SPARK-49434.

Authored-by: Herman van Hovell <herman@databricks.com>
